requirements.txt
discord.py
yfinance
flask
python-dotenv